+++ plugins/11_TheWarWithin/CHANGELOG.md
@@ -1,5 +1,6 @@
 # v85
 
+* Moved [Flamegard's Hope](https://www.wowhead.com/achievement=20594) into the achievement display group. (Hallowfall)
 * Localization updates: esES, frFR
 
 # v84